Math.ErrorProp

Description

Module ErrorProp is used to calculate error propagation in linear and non-linear systems

Documentation

data Measurement

Measurementment represented by measured value and corresponding covariance matrix

type Fn = Expr Double

um

Arguments

:: [Double]
measurement

-> [Double]
variance

-> Measurement

Measurement smart constructor: constructs uncorrelated sample

cm

Arguments

:: [Double]
measurement

-> [[Double]]
covariance matrix

-> Measurement

Measurement smart constructor: constructs correlated sample

lt

Arguments

:: [[Double]]
A matrix representing linear transformation

-> Transf

Smart constructor for linear transformation

nt

Arguments

:: [Fn]
A list of functions, one for each output parameter

-> Transf

Smart constructor of nonlinear transformation e.g. nt [x1*x1, x2, sin(x3)]

x1 :: Fn

x2 :: Fn

x3 :: Fn

x4 :: Fn

x5 :: Fn

x6 :: Fn

x7 :: Fn

x8 :: Fn

x9 :: Fn

x10 :: Fn

x11 :: Fn

transform :: Transf -> Measurement -> Measurement